+++ lib/evnt/validator.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module Evnt
+
+  ##
+  # Validator is a class used to validates params and attributes automatically.
+  ##
+  class Validator
+
+    # TODO: Work in progress...
+
+  end
+
+end