+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Version 2.0.2
 
+- Initialize code examples on repository.
+
 ## Version 2.0.1
 
 - Rename "throw" function to "stop" inside commands.