+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Initialize code examples on repository.
 - Rename the "stop" function inside commands to "err".
+- Add single validation for commands parameters.
 
 ## Version 2.0.1